WIP: add hands and blinds
ATM, a new function is function is used to generate the blind
information, but it's not test and sometimes the api fails for unknown
reasons. This need further investigation and texting. I'm pushing this
branch just to test the new strategy which make use of the new fields in
the gamestate.